Another average, set localDatetime, set longitude, set latitude, saved to db

+++ dag1/src/main/java/se/systementor/dag1/smhi/Geometry.java
@@ -11,7 +11,7 @@ import com.fasterxml.jackson.annotation.JsonProperty; // version 2.11.1
 public class Geometry {
 
     private String type;
-    private ArrayList<ArrayList<Double>> coordinates;
+    private ArrayList<ArrayList<Float>> coordinates;
 
     public String getType() {
         return type;
@@ -21,11 +21,11 @@ public class Geometry {
         this.type = type;
     }
 
-    public ArrayList<ArrayList<Double>> getCoordinates() {
+    public ArrayList<ArrayList<Float>> getCoordinates() {
         return coordinates;
     }
 
-    public void setCoordinates(ArrayList<ArrayList<Double>> coordinates) {
+    public void setCoordinates(ArrayList<ArrayList<Float>> coordinates) {
         this.coordinates = coordinates;
     }
